(feat) Add dispatch on listen to 'remove_rider' and appropriate action creator. About to work on reducers for the new functions I've made.
