+++ client/lib/sockets.js
@@ -1,5 +1,8 @@
 import { dispatch } from '../storeConfig';
 
+import * as chatActions from '../actionCreators/chat';
+import * as rideACtions from '../actionCreators/ride';
+
 /**
  *  Handle socket configuration and messaging.
 **/
@@ -16,17 +19,19 @@ socket.on('connection', function (socket) {
   socket.on('confirm_ride', confirmRide);
   socket.on('cancel_ride_request', cancelRide);
   socket.on('remove_driver', removeDriver);
-  socket.on('cancel_ride', function () {
+  socket.on('remove_rider', function (data) {
     // in-progress ride has been canceled by driver/rider
-    // reset ride state to bring user back to confirm-location page 
+    // reset ride state to bring user back to confirm-location page
+    console.log('canceling ride with data:', data);
+    dispath(rideActions.removeRider(data.riderId));
   });
 
-  socket.on('receive_message', function () {
-    // message received from
+  socket.on('receive_message', function (data) {
+    console.log('received message from server!:', data);
+    dispatch(chatActions.addMessage(data));
   });
 });
 
-// jscs: disable
 export const socketActionMiddleware =
   (socket) => (store) => (next) => (action) => {
     if (action.meta && action.meta.remote)
@@ -34,8 +39,3 @@ export const socketActionMiddleware =
 
     return next(action);
   };
-// jscs: enable
-
-function confirmRide({ rideId, driver, }) {
-
-}
